pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = ""
authors = ["zerebom <ice.choco.pudding.kokoro@gmail.comm>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.1.0"
pytorch-lightning-bolts = "^0.2.5"
torch = "^1.7.0"
hydra-core = "^1.0.4"
torchvision = "^0.8.1"
wandb = "^0.10.12"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
